--- app/src/main/java/com/example/demo/application/Task.java
@@ -1,22 +1,0 @@
-package com.example.demo.application;
-
-public class Task {
-
-    private String taskName;
-
-    public String getTaskDescription() {
-        return taskDescription;
-    }
-
-    public String getTaskName() {
-        return taskName;
-    }
-
-    private String taskDescription;
-
-    public Task(String taskName, String taskDescription) {
-        this.taskName = taskName;
-        this.taskDescription = taskDescription;
-    }
-
-}
