+++ app/src/main/java/com/example/demo/application/TaskCreator.java
@@ -1,5 +1,7 @@
 package com.example.demo.application;
 
+import com.example.demo.infrastructure.Task;
+
 public class TaskCreator {
 
     public Task create(String taskName, String taskDescription) {
