+++ app/src/main/java/com/example/demo/dto/TaskRequestDto.java
@@ -1,0 +1,14 @@
+package com.example.demo.dto;
+
+public class TaskRequestDto {
+    private String taskName;
+    private String taskDescription;
+
+    public String getTaskName() {
+        return taskName;
+    }
+
+    public String getTaskDescription() {
+        return taskDescription;
+    }
+}
