+++ app/src/main/java/com/example/demo/infrastructure/Task.java
@@ -1,0 +1,21 @@
+package com.example.demo.infrastructure;
+
+public class Task {
+
+    private String taskName;
+    private String taskDescription;
+
+    public Task(String taskName, String taskDescription) {
+        this.taskName = taskName;
+        this.taskDescription = taskDescription;
+    }
+
+    
+    public String getTaskDescription() {
+        return taskDescription;
+    }
+
+    public String getTaskName() {
+        return taskName;
+    }
+}
