+++ app/src/main/java/com/example/demo/presentation/ResourceMethodHandler.java
@@ -1,8 +1,10 @@
 package com.example.demo.presentation;
 
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+
 public abstract class ResourceMethodHandler {
 
-    public abstract String handle(String content);
+    public abstract String handle(String content) throws JsonProcessingException;
 
 }
